[roll] Roll fuchsia [build] Add new debug_symbols build API module.
Add a new build_symbols.json build API module used to list
debug symbol files and manifests for a given build configuration.

For now, the content of that files is missing:

- Entries for ELF binaries from prebuilt_package() targets.
- Entries for ELF binaries from Bazel-generated targets.

These issues will be addressed by future CLs on top of this
one. The goal is to modify artifactory to use this module
as the source of truth when deciding to upload debug symbols
to cloud storage. See associated bug for context.

This lists debug symbol files for all artifacts for the current
build configuration.

+ Add a `//build/api/client print_debug_symbols` command that
  can also resolve the build-id value for each entry, either
  using the elf_build_id_file entry, looking for .build-id/
  in the unstripped file path, or simply extracting the
  build-id value from the ELF notes section using
  //build/images/elfinfo.py.

Original-Bug: 404208759

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1236524
Original-Revision: 878a7e9e4fb9a783a7495ad26026940182035dd0
GitOrigin-RevId: e18b9b7a1ce2b1447bb0988acdda1fdbef4e7c56
Change-Id: I972e920f41a19175d812a54e1e9981990768da0e