[roll] Roll fuchsia [overnet][usb] Switch from overnet-usb to usb-vsock
Use the usb-vsock driver, which implements `fuchsia.hardware.vsock`
instead of `fuchsia.hardware.overnet` to allow the host to interact with
physical devices over usb in the same way they would if they were in an
emulator, and allow for evolution of protocols thanks to vsock having
port numbers.

This also brings in the vsock service, which does increase the size on
eng and userdebug builds a bit.

This is a reland of fxr/1232186, which was reverted due to a conflict
produced when something else had to be reverted.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238138
Size-Review: Aaron Wood <aaronwood@google.com>
Original-Revision: 978f4d2cfb6ddf39694724fe3e2b7dc0bdd3003d
GitOrigin-RevId: 00258caa0e23f99aac0bc03c65340581a6319677
Change-Id: Id7ffdc7bd359ab47819c7de90b3010629c8cf345