[roll] Roll fuchsia [vfs][rust] Remove dependency on the log crate
Now that vfs is a dylib, pulling the log crate into it forces the global
used by the log crate to exist in the dylib for all users of the vfs
crate, which affects attribution of log messages from drivers when more
than one driver using rust is loaded in the same driver host (or the
rust driver host is used).

The log crate isn't being used too much in the vfs library anyways,
though, and mostly in cases that "shouldn't happen", so this removes
uses of log from the vfs library when building for fuchsia and not for
tests.

The logging behaviour can be enabled through a new gn flag called
`vfs_rust_uses_log` that defaults to off for fuchsia that can be used
for local development purposes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235285
Size-Review: John Wittrock <wittrock@google.com>
Original-Revision: cfbeb559bbfbac0b035353aec663cca5f62920f2
GitOrigin-RevId: 3ca44cecf7e96be2a2a887ef1000fb181cfc7827
Change-Id: I84288819bb3f21007daccd4293b607ade7cc69e2