[roll] Roll fuchsia [starnix] Use zx::Counter for timers
The counter is meant to be decremented whenever an incoming message is
handled. The container can suspend anytime the counter is 0, and the
counter is incremented by the runner proxy each time it sends a message
into the container.

The counter is also set to 1 when created, to keep the system alive
until the first hanging get has been scheduled.

Original-Bug: 399939990

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238364
Original-Revision: 52ecbae92a1e6459e1b1a822b69dd9dbcefb888d
GitOrigin-RevId: a8a4c08f8065c615df1002bc2e3a981d8e50d139
Change-Id: Id19c5ece37ba53266ead58c6907ef45f7dea9e72