[roll] Roll fuchsia [power][broker] Make the ElementID an integer instead of a string
We have enough information in the events list + topology (see element
lifecycle section @ go/fuchsia-power-broker-obs-improve) to be able to
always derive the event name. Therefore, we don't need to encode the
name in the element_id and can rely on a random number for it. This
change ensures that the alive element ids are unique, however we may
reuse element ids as elements are removed which means that the event
list may have element ids that refer to different elements and tools have to take that into account.

The core goal is to reduce space usage in our inspect instrumentation.

Original-Fixed: 404631544
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232424
Original-Revision: b587b90d93313302932f04721f4f06997adf76e6
GitOrigin-RevId: a67aa6e53f8eb4cd5f38cf92f5ae20a01a4b86ff
Change-Id: I3085ea8355e22caed66e375b5fe6be91448be256