[roll] Roll fuchsia [fidl][python] Implement union, table, and zx.Handle
This change implements handles by compiling them to the int type the
same as the dynamic bindings. The complication with making this more
robust is that the fidl_codec wrapper expects all handle types to be an
integer instead of a more complex type. It would be a breaking change
for all tests that deal directly with handles to modify that assumption.

Now that all FIDL types except protocol are implemented, many
conformances tests are added so that fidlgen_python passes most of the
conformance test suite.

Original-Bug: 355717613
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1212724
Original-Revision: 4bfc6a4de975977c631e2e16e4357019403e1ba3
GitOrigin-RevId: 0056cb255613565ee3b41e95d6fc4639ecb55669
Change-Id: I3aa735641bbb1375d5145f17dc4c48e73cc6dfb7