[roll] Roll fuchsia [kernel][vm] Remove freeing_owned_pages
Existing call sites sometimes have to 'lie' when freeing pages from a
hierarchy and claiming that the pages are not owned by this node. While
true, they are owned by some node, and should logically be freed by
that node. The argument put forward in these locations is that
hierarchies do not need to be correct, since they will never have a
page source handling free.

Instead of having call sites make this argument and pass a false value
for freeing_owned_pages, this change puts it forward as a global
property, that objects that have a page source handling free do not
have children, that can be checked and then generally assumed. Doing
this allows for removal of the flag passed to free and simplifies
call sites where it was otherwise ambiguous what 'ownership' meant.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1236944
Original-Revision: 4ece84db3f908c9943696039a210cae12e6bc0fd
GitOrigin-RevId: fe26f26d61ae945f92224148b22a92baa306f4df
Change-Id: I4008a32feb32c8298445baf3b0a47863f3919348