[roll] Roll fuchsia [fdf][rust] Introduce `OnDispatcher` trait
This trait allows for a uniform abraction of a dispatcher that may or
may not be available when it comes time to submit a task to it. This
allows for using a 'Weak<Dispatcher>' directly with channel reads and
task spawning, but also will allow for using the "currently active
dispatcher".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1223306
Original-Revision: 3cd3a7cd597ac9c9fd405771d6958402c631dc58
GitOrigin-RevId: 1532bf87df502379932cf1d737c1a3478dda9859
Change-Id: I00e647eba92e8c76e14581942baa92039cd4d48f