[roll] Roll fuchsia [netstack3] Send ICMP Port Unreachable from UDP
Originally, Port Unreachable messages were not sent by the netstack. The
rationale was to avoid contributing to a reflection DOS attack, in which
the attack performs a UDP port scan on the netstack, causing us to send
a large volume of ICMP Port Unreachable errors to the victim.

This attack vector is not a substantial concern, because the netstack's
ICMP implementation enforces a rate limit on the amount of ICMP errors
that can be sent. This is in line with the behavior of other platforms.

Original-Bug: 406666070
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1239544
Original-Revision: a452e24d9eae9a259e3e5eed540668d65829f61c
GitOrigin-RevId: 0d2bde6c760bd8879eb440d7eeb2786f3ac2b359
Change-Id: I83e6b78a0f85a135722092274f5473c5b388382f