[roll] Roll fuchsia [netstack3] Update MSS on PMTU update
Respond to PMTU updates (delivered as ICMP errors to TCP sockets) by
adjusting the sender MSS (and congestion window) accordingly, as well as
eagerly retransmitting up to a single MSS starting at the segment that
was reported to be too big.

Eagerly retransmitting the segment has the advantage of allowing PMTUD
to continue right away, which may prove useful if there are other MTU
adjustments that need to be discovered along the path to the remote.
Note, however, that we *only* retransmit the one segment that was marked
as too big in order to avoid causing a packet storm in this case. Any
other data that must be retransmitted is handled by the retransmission
timer.

Original-Fixed: 404628798
Test: netstack3-core-tcp-test
Test: netstack-socket-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233389
Original-Revision: 34ea50ba339ea028c8912da3d8a8bd2c22cadb3e
GitOrigin-RevId: d417936a1ccf7ec1d587907b6ad83d09b890c999
Change-Id: Ie7bc16b68e8461b2be153775b4cc95ed7da8624b