[roll] Roll fuchsia [bazel] require target name in visibility_templates
When a user uses wild card expansion in the visibility_templates
attribute on the fuchsia_sdk_repository rule we need to ensure that they
have specified a target. If they add the '*' at the end of the label it
will try to use '*' as the target name which is not what we want.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1240526
Original-Revision: 291f9fe5484f3315b2e2f57cca90c0997f96aa5f
GitOrigin-RevId: aadef67ced1353d54e113d2324249c3a240cdce9
Change-Id: I3e76e9417cccdef1c2563d60c38c916548ceca10