[roll] Roll fuchsia [netstack3] Limit size of PMTU cache
Use an LRU cache to limit the number of PMTU entries we maintain. This
prevents a denial-of-service attack where PacketTooBig messages are
repeatedly sent to the stack until it OOMs. We discard the
least-recently-used entry when the cache becomes full; currently this
will just be the entry that was inserted first, since the cache is not
actually *used* for IP fragmentation yet. This property will be more
useful when that is the case (see https://fxbug.dev/404629697) since it
will prioritize in-use paths to remain in the cache.

Original-Fixed: 404634511
Test: netstack3-core-ip-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1237566
Original-Revision: e6ce23cd0cd732fc3ac81d619a053d528b79aff1
GitOrigin-RevId: df56c8c7729060d2875686319f76b5e6039e546e
Change-Id: Ic1756c1c6deeac5c777195a98a60c939d3c5f09a