[roll] Roll fuchsia [fuchsia-controller] Only allow unions with exactly one member
The dynamic Python FIDL bindings for union types enforce only one member
per union, and the variant must be specified upfront (except in the case
of a flexible union with no variant, or when using make_default internally).

The primary motivation for this change was a bug in the implementation
of fuchsia-controller bindings for FIDL methods that take a union for
their arguments. The implementation required positional arguments for
the method, but the union type did not accept keyword arguments.

There is a known bug that still exists even after this change which is
that the comparison of two flexible unions with an unknown variant may
return a nonsense boolean.

Test: Add tests for unions in static and dynamic bindings
Original-Bug: 404285816
Original-Fixed: 42078357
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1223484
Original-Revision: 11fba0499c66800fa542ba57df278f263e0ed3f4
GitOrigin-RevId: f26fdf1a6fbdbfb1641a7d3f2dbdc857a78095d9
Change-Id: I6cbbfbd4b3c4f0c1e078c28c64bab78d8a5eb69b