[roll] Roll fuchsia [fidl][python][fuchsia-controller] Implement FIDL result unwrap for static bindings
Each FIDL result is a union of at most three variants: response,
framework_err, and err. Each variant could be None, and response can
even be None when it's present. The dynamic FIDL bindings have an
unwrap() method to conveniently extract the response field or raise an
exception if it's not present.

This change implements a similar unwrap() method in the static
bindings, and modifies the unwrap() method in the dynamic bindings to
match. This modification is to check framework_err if it's part of the
result type.

There is however still a difference between the two implementations.
Because the static bindings do not compile empty success structs,
unwrap() can return None when the response is an empty success struct.
In the dynamic bindings, unwrap() would return the empty success struct.

Test: Add many tests for unwrapping a FIDL result
Original-Bug: 355717613
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238164
Original-Revision: 7aa6942b6d49a91f2b894dc46c4170242c47b4d7
GitOrigin-RevId: 4c07baa8af4d6b83f15e82f7aef25b5f790d5639
Change-Id: I4599d715454f52108bbe44cfa34cd615da818f35