[roll] Roll fuchsia [kernel][vm] Refactor RangeChangeUpdateCowChildren
Makes RangeChangeUpdateCowChildren a static method that takes a
LockedPtr, instead of adopting a Guard. Using a LockedPtr is simpler
for the final implementation and does not impact call sites.

As this method does not have public callers it is moved to the private
section, where the LockedPtr is declared.

Original-Bug: 338300943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1236964
Original-Revision: 1933a87c35ab2a4360bff4f35b43c7250397bce2
GitOrigin-RevId: 53130802e0be77d86c66fb4af99655a90ef2eb44
Change-Id: I2c8856f853f697356aaa8976ae20f734b1d4b996