[roll] Roll infra/recipes [tricium] Use git_checkout for non-fuchsia checkouts
Every place that we use Tricium for a non-fuchsia checkout it's checking
out a single repository, so there's no need to use Jiri; we can use pure
Git.

This should also fix the breakage of the non-fuchsia.git tricium
builders caused by Jiri change https://fxrev.dev/1239744 which started
enforcing that the `jiri import -name ...` argument corresponds to the
name of the project with the same remote, which is not the case here
because the tricium recipe doesn't support setting the import name for
non-fuchsia.git checkouts.

Also delete the `manifest/minimal` manifest because Tricium was the only
thing using it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1241604
Original-Revision: fe725de955f752dc3bf65c9736bd4df63e50a715
GitOrigin-RevId: 95d6196c330067a0a5f834b5172068d90b86d52f
Change-Id: Iefbafb55b9348ebff53791d83f974564f3e8ccf5