[roll] Roll fuchsia [owners] Remove dhobsd@google.com from src/connectivity/network/netlink/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/dhobsd@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241790
Original-Revision: 6cd535b045130887d8e6ff731d40547d908e8150
GitOrigin-RevId: cd16901a0e13dcedf53bb0a27a6245662a246974
Change-Id: I7286d1d6884126550f56a39b0eb2a6cae9759ca0