[roll] Roll fuchsia [owners] Remove dhobsd@google.com from build/go/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/dhobsd@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241789
Original-Revision: c08811ace379097ee4e61b03f5b72d70f134ef9c
GitOrigin-RevId: f4289d82f7c5a9125d407ac1a54de3f65be4f609
Change-Id: I340a90a23b8cf34ef043c6d35125e50ec479486a