[roll] Roll fuchsia [test_pilot] add summaries
This CL adds summary processing to test-pilot. In addition to creating
a default summary for tests that don't produce one, this CL makes
test-pilot merge summaries from multiple sources.

The actual names of directories used in the output directory and the
way we deal with summaries is likely to change as we integrate
test-pilot into the test architecture and finalize the output format.

Test: new unit tests

Original-Bug: 327640651
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1240607
Original-Revision: 90082828da05c822a43547a4df8e53a13c82bbdb
GitOrigin-RevId: 3851ed8d38d3d9e3faa4a69703fd42e343bb03f5
Change-Id: I38fdf3df182886a18dd3bb5e844c5ae039816ffa