[roll] Roll antlion [owners] Remove dhobsd@google.com from packages/antlion/controllers/fuchsia_lib/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/dhobsd@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/1241788
Original-Revision: 88745619663c1aa7eacec787a33bd54105cec629
GitOrigin-RevId: a18f910fe21a4b2c32c254938f5406e25e974973
Change-Id: I280f804c28ead820c430382922b3bb940becca97