[roll] Roll fuchsia [fxfs] Don't use the root_gen_dir for unicode_gen
This results in rust-analyzer thinking that the set of possible
dependencies for unicode_gen is all of the root_gen_dir, which is
thousands and thousands of files.

Instead, put unicode_gen.rs in its own directory, which lowers the
number of dependent files for the unicode crate from 45324 to just a
handful on a minimal.x64 build.

Test: build still works, existing tests.
Original-Bug: fxbug.dev/270999861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1242964
Original-Revision: 43d76062b6c8ceee835f4de269821d0ca18e2c88
GitOrigin-RevId: 08d2bc16b47f98386d72622e1bbe4e8b8cd122f1
Change-Id: I536c31d21ee633e54d22171f9d931d9d5a9f6735