[roll] Roll fuchsia [assembly] Allow clobbering dtbo from different assemblies
If the main images and recovery images both add a DTBO via the board
and if recovery uses AB-slotting, then two DTBOs will be destined
for the same A parittion. We allow this as long as the DTBO images
are equivalent.

As DTBOs are small (a few MiB), then hashing the contents is very
fast.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241729
Original-Revision: 8a56c8141a02f38968abbb881f5d14b845007b82
GitOrigin-RevId: 9f8836e59b0147bb7744bab6e23f57e0ffd3510b
Change-Id: I5fddc1986394c73f5d556eecbf6f167341940359