[roll] Roll fuchsia [usb-rs] Manage buffer lifecycles correctly
The previous code unsafely leaked pointers to the reaper thread for the
buffers. Now URBs each have their own associated buffer which is tied to
their lifecycle.

Also we now try to cancel URBs when the waiting future is dropped.

There's still a chance that the reaper thread could successfully reap a
URB but the future could be gone already, in which case the data in the
buffer will be lost. So there's one more potential bug to squash.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1236764
Original-Revision: 5a3821f861d71a4b776555bce0a4014af2a644de
GitOrigin-RevId: 0e4c74622095c7e8d189f100fc846f63fccc7ad2
Change-Id: I158d76940a64a51cd6a2f0e09a6cf52a069d50b0