[roll] Roll fuchsia [sdk][fidl][gpio] Define WAKEABLE interrupt option.
Previously we passed through the ZX_INTERRUPT_WAKE_VECTOR kernel flag
to syscalls, but this didn't work for all pinimpl drivers. For
example, some pinimpl drivers demux the wakeable interrupt to virtual
interrupts, where the kernel flag should not be passed through.

This change defines a bit "WAKEABLE" which is not tied to the kernel
flag. An interrupt is "WAKEABLE" if it can wakes up the system if the
system is suspended.

Original-Bug: 361851116
Original-Bug: 380120239
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1240887
Original-Revision: c23ef38214e2c3c7ad39e1e7ed1c0014725ac516
GitOrigin-RevId: e86b6a031139c9deb69965ac35c126e8251c0062
Change-Id: Ie0404e52b619e8fbe5347a53ef02cc34ae5b10c3