[roll] Roll fuchsia [ffx][usb] Add echo packets
This adds echo packets to the usb vsock implementation library and
automatically replies to them. This will be used for probing the usb
connection at synchronization, but might be used for other checks to
validate that the usb connection is still in a functional state.

Original-Bug: 406262417
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1245327
Original-Revision: 1fb9b6e50076f7281117f61d3c90f08fe7981972
GitOrigin-RevId: c7895da0eae85ef987a6f2bc7010f03f0ac56f9e
Change-Id: Ice0f5376997479a2ae1e891d2c8e306bf3417807