[roll] Roll fuchsia [fdomain] Check channel message size before write
We want to bound the FIDL vec size for the write call, but this will
panic in the client if the message is too big rather than sending the
message and letting Zircon reply with an error. So we add a manual
bounds check and produce the correct error synthetically instead.

Original-Fixed: b/388093085
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217866
Original-Revision: 48f7ea5ba48a97619b86d69873093d0279ac362b
GitOrigin-RevId: b0d10f1c45a924bedebc9f99528049d7980bcdbf
Change-Id: Ib0872ae03c282010c06b1039ed8c4839f0100940