[roll] Roll fuchsia [build][api] Fix --last-build-only implementation.
Fix the logic used to implement `--last-build-only` by:

- Using the new Ninja `outputs` tool to get the set of artifacts
  from the last build invocation.

- Use the `inputs` tool to get the set of sources used by the
  last build invocation.

- Use both sets to filter build API modules. This ensures that
  both prebuilts (sources) and intermediate artifacts (e.g.
  unstripped binaries) are correctly listed.

NOTE: The bug is not fixed because prebuilt binaries, like Clang
and Rust runtime files are never part of the result. The reason is
that they are not listed as proper implicit inputs by Ninja targets.
This will be solved in a separate CL that adds the proper data_deps
to C++ and Rust binary targets.

Original-Bug: 408504040
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1246784
Original-Revision: d3778db2876708724904363d29c35e114975fb13
GitOrigin-RevId: a68c885f5c3fd05c40ed67acc9ac72f48a5bcdd2
Change-Id: I3e145e1c5aa5882d3e0e46ca0d3607546ae48547