[roll] Roll fuchsia [kernel][vm] Refactor loaned page replacement
Moves the loaned page replacement in ProtectRangeFromReclamationLocked
to be outside the main lock acquisition. This is necessary since the
loaned page is being replaced on a potentially different VMO to what is
otherwise being operated on and so will need a different lock
acquisition and DeferredOps.

Original-Bug: 338300943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1242585
Original-Revision: b2b2a849113650fd74c36d6d5eb6acce2fe06c30
GitOrigin-RevId: 9029b3a95116bc6b6bde4caacc4ab96e519b3069
Change-Id: I0a8b8f2eb7324c75ade084b223a3c89e4747a7f4