[roll] Roll fuchsia [kernel] Add Drain method to SpscBuffer
KTrace allows users to "Rewind" the trace buffer, which effectively
means emptying it. This CL adds support for this functionality to
the SpscBuffer with a method called Drain, which is logically
equivalent to performing a Read into /dev/null. Callers could have
already implemented this function themselves by calling Read in a
loop with a no-op CopyOutFn; adding this to the API just makes the
operation more efficient.

Original-Bug: 404539312
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1237565
Original-Revision: 49136bdd3cd1073009cedf5cb09aa85fbd821daf
GitOrigin-RevId: 0ddf4046a3e1b031656de3d0c72ff8f16866474e
Change-Id: Ice7683a4a4c2a0e16d23620e9f9c24afb79b1c0d