[roll] Roll fuchsia [ffx][log] Adding symbolizer as a data_dep
This adds //tools/symbolizer as a data_deps for the `ffx log` subtool.
The subtool invokes `symbolizer` and if it is not built and installed as
a host_tool, ffx cannot find it in-tree.

Out of tree, the sdk manifest is used to locate it.

Original-Bug: 409133313
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1248005
Original-Revision: b24278d7ac3c60b48e85c922127d7ddad692c0c4
GitOrigin-RevId: 4139f915c3ee7a9a2108864f725ba3da33c3accc
Change-Id: I710571f212c86d8b525338d99bf8123aeb2d215d