[roll] Roll fuchsia [display][coordinator] Migrate integration test to Virtcon fallback.
This CL updates IntegrationTest.SendVsyncsAfterImagelessConfig to wait
for the virtcon client to gain display ownership after it connects to
the Display Coordinator. This ensures the client has ownership before it
calls ApplyConfig(), which is the path we intend to test.

Original-Bug: 388885807
Test: fx test //src/graphics/display/drivers/coordinator
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1244566
Original-Revision: f1e3d55be14e476ea751dd886e8e17ade582451e
GitOrigin-RevId: d19b063f18c425450024b5446b35fec206194554
Change-Id: I9ae9028f43cd0a5925aee036932fac118343dae1