[roll] Roll fuchsia [archivist] Add e2e benchmark
These two benchmarks validate the time it takes from emitting a log
until we read it and parse it back from archivist. It's possible we may
want to remove the parsing side of it, but leaving it there for now,
although it may inflate the numbers a bit.

Original-Bug: 300156427
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1247304
Original-Revision: e75f017c6b7e6f7bada616a392f1f6b20a316eaa
GitOrigin-RevId: f3e4924b4d55764987f1ad17d9b17bb81fb47d0e
Change-Id: I01136f0bf74ea31f9af6130f96de1f9f595ee5b1