[roll] Roll fuchsia [wlandevicemonitor] Lock inside destroy_iface.
Prior to this change, it was possible for a future
communicating with a PHY driver to destroy an interface
and a future monitoring the interface for liveness to race
to complete.  When the monitoring future won the race, the
rest of the system was notified that an interface had been
removed.  If another component attempted to create an
interface at this point, the operation would fail because
the destruction process was still ongoing in the PHY driver.

This change introduces an async mutexing scheme around
calls to destroy_iface.  In the case where an interface is
being destroyed by an API request, listeners will not be
notified of interface removal until the driver has
responded to the destruction request.

Original-Bug: 370381717
Test: fx test wlandevicemonitor-tests
Test: fx test hw-sim-tests
Multiply: wlandevicemonitor-tests
Multiply: wlan_policy_ap_test_fc: 500
Multiply: wlan_policy_test_fc: 500
Multiply: wlan_test_fc: 500
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238716
Original-Revision: da8bb481809b0e0fa4b0e8dd77cd719fdf279d79
GitOrigin-RevId: 13199a064295e14e75b711133ed0b1d1fee9e1aa
Change-Id: I255efa9c5ef0eefcf7c6ae1ca9a1828a5e6ed0e6