[roll] Roll fuchsia [sherlock] remove gpt type guid remapping
This metadata block would tell the gpt driver to remap partitions with
certain labels to have specified type guids. The two consumers of
partition metadata are fshost and the paver. Fshost doesn't use type
guids for identifying partitions.

The paver does, mostly via FindPartition, but there is an issue with the
implementation - FindPartition reads the gpt metadata off the disk so it
ignores the remapping (which is at the driver level), but then it passes
the non-remapped type guid into OpenBlockPartition, which _does_ use the
remapped type guid to find the partition and open a connection.
Therefore, if the remapping applied to any partitions in use, OTAs would
already be broken for sherlock. Since they aren't, we can assume the
remapping is a no-op.

See the bug for a slightly more detailed discussion.

Original-Bug: 393194863
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238668
Original-Revision: 500a41a1e4aed48d1035d6f73a2612dd959043f4
GitOrigin-RevId: 3533cacafb7c8a828374030304632d9d5add444b
Change-Id: I05104bc49115b64793c754361c8b7cea221059f8