[roll] Roll fuchsia [netstack3] Reset dev params on interface disabled
When an interface is disabled, reset it's learned IPv6 network
parameters (i.e. the retransmit timer learned from NDP).

When the device is re-enabled in the future, it may be on a different
network, and any previously learned parameters won't be applicable.

The IPV6-AUTOCONF network conformance suite relies on this behavior to
reinitialize interfaces between cases. With this change, the sequence
"2.2, 2.3, 5.1" is no longer causing cross-test-contamination.

Original-Fixed: 407754539
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1246645
Original-Revision: e88a6be3fd22be98ceeffde8fcf98831b8861a45
GitOrigin-RevId: 79a77b7b0d8a3d999030b563151aac2e4d5229f1
Change-Id: I09a2726a93661a888d4bdf713301719f436ad121