[roll] Roll fuchsia [starnix] Consume all ComponentController requests
We were only consuming the first ComponentController request in the
stream, which means that if a Stop request came in, and then timed out,
and subsequently a Kill request came in, we'd never observe the latter.

Address this by consuming the requests in a loop so that even if we've
already started shutdown due to a Stop request, we still respect
subsequent Kill requests.

Original-Bug: 409364939
Test: _session_restart_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1249005
Original-Revision: 3f7cece5a1a9ab048009ddcd5cc71f4674a9ff82
GitOrigin-RevId: 27af6970bb99986133ed45c69e36882d5c372334
Change-Id: I836702af390e272d289fd93db7fd0434b4e5588b