[roll] Roll infra/recipes [cache] Don't override errors if purging fails
If an exception happens within the cache context, but purging the cache
fails, still propagate the original exception to the build's
summary_markdown instead of overriding it with the cache purge
exception.

It's not the end of the world if the cache purge fails since it will get
cleaned up at the start of the next build due to the guardfile still
existing, and certainly not worth overriding the actual root failure
reason.

With the old behavior, the cache purge failure would mask the original
error. This was rare, but it did happen if there was a timeout of the
context wrapping the guard clause, which would first cause the wrapped
step to fail and then cause the cache purging to fail. For example:
https://screenshot.googleplex.com/A544hB4jsaQq5Q8

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1249305
Original-Revision: 41041d08de51765ee7942770b7f34650ca06a06b
GitOrigin-RevId: eea76680ed9cf6e9a42a163973fcad8dff965c62
Change-Id: I71d7e18b7923a7d5b0456762adfc5f9b4bc77d88