[roll] Roll fuchsia [persistence] Use B-Tree to group tags by time
Refactor Scheduler to use a B-Tree while grouping tags that need to be
fetched in the future. This is meant to be a readability improvement.
While this also improves over the existing O(N^2) solution with
O(N log N), it may not result in better performance for small values of
N due to the constant time for creating a B-Tree.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1247458
Original-Revision: e14555866daf2363f33265d07d6efd5023fa4679
GitOrigin-RevId: 99b04f4c4d0a3849d101c51482c1ca7e3a5be31f
Change-Id: I7ba22b173312e05d8935b8dd27e4b0e0067be474