[roll] Roll fuchsia [fshost] Fix paver shutdown on qemu
On storage-host configurations, the paver opens PartitionService in its
service directory to enumerate GPT partitions.  Fshost hold the server
end of this service in escrow until the GPT is found and bound to the
GPT component.

That might never happen, since some systems don't have a GPT.  For
example, an emulator launched without the `--uefi` flag doesn't have a
GPT.  In this case, the open call to PartitionService would hang
indefinitely, which means the paver will be unresponsive.  Since we
can't OTA on such a system, this only manifests during shutdown.

To address this, if fshost is still buffering PartitionService requests
when the Fxfs container is found, fshost will drop the buffered requests
(and clients will get a PEER_CLOSED), unblocking clients and informing
them that they cannot access the GPT because it doesn't exist.

Note that the only other system that doesn't have a GPT (astro) is not
relevant here, because in that case the AstroDevicePartitioner never
bothers to attempt to find the GPT at all.

Original-Fixed: 408423972
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1249746
Original-Revision: 5039f3ac5e0c8c9a4c4f7bd3453828eb50fcceea
GitOrigin-RevId: 7b53ae36aa7db5576ae5bd0d2d3f28eefc42e1bf
Change-Id: I7e73da12ce580ee352e99f603e7b51d1119dc699