[roll] Roll fuchsia [starnix] Fix race in thread and task release
When a Task was released, its WeakRef could not be updated anymore,
while the TheadGroup that was owning it was not updated yet and could
still be accessed through the pid table. This ended up creating
ThreadGroup with no available task in.
This change the Task release to require a lock to the pid table. It
ensures that while holding the lock to the pid table, any thread group
read from it will have at least one active task.

Original-Fixed: b/409478205, b/409425986

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1247587
Original-Revision: 0aff63b1bd53835279050b68fc14d7299d033095
GitOrigin-RevId: 28a23043e0a9934a2dd06f33c17787ae39940112
Change-Id: I0a0b1e285bfd7d28f4f0ccd7d94e3114ae4bfa18