[roll] Roll fuchsia [fxfs] Cache errors when replaying profiles and downgrade log level
This gets noisy for something that is actually quite acceptable. It was
originally WARN level on the assumption that after an OTA none of the
files should actually have gone missing yet, but that can be racy with
whatever the SWD stack might be up to, so it might not indicate an
actual problem.

Adds an info log to indicate count of success and failure, so if this
becomes suspiciously high we'll have an indicator without creating a lot
of log spam.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1249144
Original-Revision: 48b74a4f7ccffa941ef08f627b2a6bf01abfdf44
GitOrigin-RevId: 4b1b6013ecb2aa5dad9d443a827db05100d512e5
Change-Id: Ib79a96441768f05bcac22337ee71f63ef9e91a0c