[roll] Roll fuchsia [kernel][vm] Refactor PageFaultLocked
Restructures the PageFaultLocked code to first cast to the underlying
VMO type, and then take a lock on that object. This allows for the
VmObjectPaged case to create a DeferredOps outside the lock, but only
if it's a paged object. This requires some small duplication of logic
inside each of the if/else branches.

Original-Bug: 338300943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1242596
Original-Revision: 567d50bd76944772d9c3cf74d35ed355955dc8b4
GitOrigin-RevId: 4d6a6308a40d72b33f00cc9feefb45d695588742
Change-Id: Ic86c18b528cc9c2f7a6a2e0a51ab3089e7dba834