[roll] Roll fuchsia [blobfs] Check if the preexisting BlobWriter is still active.
Verifies that a preexisting BlobWriter is actually still active by
calling `signal_peer()` on the channel to see if the client side has
closed but the executor just hasn't handled the message yet. This
provides protection against racing with the old client handle close
and the new BlobWriter creation.

Adds bare pointers as back-refs from the associated blobs to their
BlobWriters. This lets duplicate blobs found in the cache to allow
manipulation of the BlobWriter. In particular, it allows dropping the
Blob so that it gets recycled out of the cache and makes room for the
new version being created.

Original-Bug: 408439508
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1250905
Original-Revision: 67c8298249c0d816e5a757f2e2d1724d5e16af3b
GitOrigin-RevId: b5967b0a73dbac155fdbf683c0bd5987b578d80e
Change-Id: Id4ba696ca83f14671227ad49f2ae92813285b5de