[roll] Roll fuchsia [symbolizer] Change log level to debug for symbolizer
I made this change because when `ffx profiler symbolize` be called, a
lot of non-fatal errors been print out to the screen for the profiler
users. A lot of backtraces be handled at once, and of course, some of
them might fail, and users are not care about the detailed failure
reason.

Original-Bug: 409569703

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1251015
Original-Revision: 422bb9fa749bd70f0df4357b3b0ead37b269a386
GitOrigin-RevId: e3e3752cd95d24220ebcec0083ee8ad9af4fcf7e
Change-Id: I1dab2ed5093c0f8002410fd8843820c4f1ebc8df