[roll] Roll fuchsia [ffx][discovery] fix dangling mDNS query loop
As opposed to `Task::local(...).detach();`, now keep track of the
recv tasks for IpV4 and Ipv6 loops.

Note: `detach()` will cause the thread to not be stopped when the task
is dropped.

This will stop the Tasks (and therefore the dangling mdns traffic) that
the mdns_discovery lib was creating.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253433
Original-Revision: f583363dcbfe08cf92db6167db52e3164c12628d
GitOrigin-RevId: 56e9dbd632eb028778d7315f7e64397903cfae1c
Change-Id: I47519fa17cd5c57ef70fdc35bbb8f5cb348bf938