[roll] Roll infra/recipes [gds] Simplify + clean up tests
While working on a change that touched this recipe module
(https://fxrev.dev/1253666) I noticed that the tests were super fragile
because they depended on the order in which footers were read. In
addition, it was very difficult to debug the tests because the test
cases in full.py were separate from the test data in test_api.py.

So I made a few changes:
1. Delete the separate change_details request for each change to get its
   dependencies - instead, use the appropriate query params to get the
   commit messages when listing candidate changes, cutting out an
   unnecessary gerrit request and eliminating the need to set step data
   for that step.
2. Move all the test helper functions into full.py where they're used,
   so it's easier to inspect the test data.
3. Use fewer defaults in test data, require more constants to make it
   easier to understand the assertions the tests are making.
4. Make test data more readable, e.g. "I12345" and "I67890" for test
   change IDs to make them more easily distinguishable versus "I542243"
   and "I543343".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1254144
Original-Revision: 721db96932e96804f73d7e3faedc48052956f195
GitOrigin-RevId: 48e157e0c1bd2c4daf7928a5eb6258751c4da10f
Change-Id: I0f426221f9365b91db364308748a0ae0742b708d