[roll] Roll fuchsia [thread-sampler] Add asserts for per cpu structs
Timer migration could be a possible cause of the panic we see in the
thread sampler code. We don't expect timer migration to run in a normal
profiling scenario, but as the sampler code is currently incompatible
with timer migration, we should at least assert some invariants to
ensure it doesn't happen.

Original-Bug: 410034543
Original-Bug: 392888672

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1250734
Original-Revision: 169cae3dbdb9fb86eef2778aea46ecab1a6a2160
GitOrigin-RevId: 53e583e02e3f1dea76e6a590833c11335f7f3543
Change-Id: I2c2294aac75e1da6f37ce771abe084a0b47dd26a