[roll] Roll fuchsia [fidl][rust_next] initial GIDL conformance tests
This is a work-in-progress GIDL backend for the new Rust bindings.

So far we have most encode success cases building (the failures are
deny-listed away) and most of the building tests passing (315 of 355).
The failing tests are not deny-listed but the rest is not part of the
default build.

Nothing with handles works yet. Decoding, failure testing and
benchmarking doesn't work yet.

The rust_next GIDL backend is based on the rust GIDL backend and still
has a bunch of unused, outdated code to be ported as we complete it, but
it's useful to have there are a bit of a path to follow.

The tests can be built and run with:
  fx add-test //src/lib/fidl/rust_next/conformance:fidl_rust_next_conformance_tests
  fx test //src/lib/fidl/rust_next/conformance:fidl_rust_next_conformance_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1249213
Original-Revision: 96d5dc4cba2d110a397ff952ce00a6ae107f9112
GitOrigin-RevId: 9dcffe48144a42a4fa9eb66caef1cae1c7179aff
Change-Id: I12c2bc692fdcc91613c37f4a2de515a62f53098c