[roll] Roll fuchsia [trace][starnix] Fixing ATrace End parsing
According to the Perfetto systrace parser code, it is allowable to
have an End event for a thread scoped event that does not contain the
TGID. The reasoning behind this is since it is a thread scoped, the id
is not needed to match up with the Begin event.

We currently use the pid to manage the event queue, so in the case of
End missing the TGID, the current task pid is read to find the correct
queue.

Original-Bug: 401645216
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253428
Original-Revision: 216ab999bd5af4c5d79f5380d60fde55d8703941
GitOrigin-RevId: d9a692f7d51b0b8a5c67f903922a790f5b297fbb
Change-Id: I34eb638ec69e1249367a316052e4373b012b91e5