[roll] Roll fuchsia [graphics][drivers] Add a TestService
Update the test drivers to expose a unique service that gives
access to the test protocols. This means that we can be honest
in the CML and FIDL type system about what capabilities the
drivers are exposing, and we can remove the templated FIDL
protocol serving from the drivers.

This is a soft transition as OOT drivers will need to be updated
before the old functions can be fully deleted.

Original-Bug: b/410611164

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1252545
Original-Revision: 7b242bfb9398c18ac9b15c558fa5880dd6fdee32
GitOrigin-RevId: 42c790f68c114491baa113402f5bd7acce1154ea
Change-Id: I77225a000c04c7a06b22dbf034f6ae4b81454ecf