[roll] Roll fuchsia [wlan][policy] Use global backoff, stationary roam
This change updates the roam scan backoff logic in the stationary roam
profile to use a simpler global backoff. This is because, at release,
there will only be a single roam reason, so much of the existing backoff
(e.g. extended only when there are not new roam reasons) is irrelevant.

Eventually, we may need a backoff solution that tracks multiple reasons
again, but there are some edge cases to consider. That effort is tracked
in the bug below.

Original-Bug: b/409839414
Test: wlancfg tests update to reflect new backoffs

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1254547
Original-Revision: 5e0e2b986a3237130ce698e85532230a4371e7d5
GitOrigin-RevId: db318654e194864aca1c3971b7f0f90919f9ff4d
Change-Id: I0c5e6449f3620c4a69beb8738656615f291801af