[roll] Roll fuchsia [fidl][rust] Re-enable protocol tests
These client-side tests were previously disabled pending diagnosis of
why they were failing. They were failing because the client sender
needed to be dropped in order for the channel to be closed. This isn't
desirable, but for now I've fixed the tests with a TODO and re-enabled
the tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1254286
Original-Revision: b9964bb49332fba7e955e356643c33317f9828a6
GitOrigin-RevId: a249de5a68ba741ebfc573fa8c85418076b80234
Change-Id: I4de840a1157661de6af3224fde1a110e0b1a2d66