[roll] Roll fuchsia [packet-formats] Allow updating raw ports/addrs
As part of this, update the header checksum (if present). We don't know
if the checksum is valid, but we at least maintain that any updates
don't make it *more* invalid :-)

This will be used as part of NATing ICMP errors, where we want the inner
packet to have the "right" checksum even though we might not have the
whole packet in order to validate it.

Original-Bug: 341340810
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1255405
Original-Revision: bb0623cb279074d6dd38696647cd69c71c1f36b0
GitOrigin-RevId: 35e4d9f31276dc89f9276e470dfd95283cec622e
Change-Id: I03ee4b5071740a41c3ae0650f789207bd157e042