[roll] Roll fuchsia [ffx] Make new target list JSON schema compatible with old
The introduction of non-IP address types changed this schema, and we
keep tripping over migration pains. This changes the schema again to
something that looks more like the old one and should let tools ignore
it if they haven't been updated.

The places where we attempted to migrate all had to keep backwards
compatibility with the old schema so this should be able to land safely.

In the future we should find a way to make changes to the schema goldens
block changes harder. As it stands they're just golden files that in a
lot of engineers' build configurations will just automatically update if
you change the schema.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1255852
Original-Revision: c8bd8d7dfec5f625e6369dadcc7209684cd11106
GitOrigin-RevId: 5f6fd099f5656d687ee8f18d9bad2d0be2b12c1c
Change-Id: I0b41d08cb8c6c4e9747135194252468d88a0d0e1