[roll] Roll infra/recipes [fuchsia_roller] Break up multi-commit rolls into multiple CLs
If an incoming roll contains multiple commits and the roller has
force-submit enabled, split up the roll into a separate CL for each
rolled commit (up to a limit).

This is something I've been wanting to do for a while anyway, so that if
a force-submit roller fails due to a transient failure mode it doesn't
squash CLs together in the next attempt, which would make blamelists
harder to use.

It will become more important when we combine the fuchsia.git and
vendor/google rollers into one roller that can roll the two repos
together as needed for circularly dependent cross-repo changes. We want
the roller to be able to isolate the circularly dependent changes into a
single commit with no other extraneous commits included.

Original-Bug: 410922525
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1249304
Original-Revision: e6967285dc717d4225ad1a37f909522d066985d7
GitOrigin-RevId: e8295b4fb4b277d63d87cc97b299f1678fc73c78
Change-Id: Iad2a2d1762c085270988c4677b00b84ddc9e8979