[roll] Roll fuchsia [fidl][rust] Escape hatch for non-blocking sends
There are a few places in the storage benchmarks where we assume that
FIDL sends complete immediately. We prefer that users await their sends
so we can add backpressure support to channels in the future, but having
an escape hatch is useful for porting.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253364
Original-Revision: ddc60f5acf405dc174de73974e0bfd4241e47b7b
GitOrigin-RevId: 088d0f900fee01d75c87594a2df89086a057f1f5
Change-Id: I3eea47171c3e1c2b7acb0320ef22dc727dbd8bd5