[roll] Roll fuchsia [netdevice] Fix crash when switching sessions
Observed when adding controlled shutdown to netstack3. There's a
specific series of events that cause the attempt to take over primary
session when a session newly attaches a port that attempts to insert a
nullptr in a doubly linked list.

Fix the code with a simple check and add a regression test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1259365
Original-Revision: 55a0c3fb6555295c89dc5569179368fdab1a1f5a
GitOrigin-RevId: 20d1ed1f0656c93daf00bff3431a07592daa7a95
Change-Id: I30f71181343a3d9990ed010da848b48e48da5c03