[roll] Roll fuchsia [netemul] Enforce synchronous checked shutdown
Change the netemul crate to, by default, enforce a synchronous checked
realm shutdown whenever we drop the realm handle. This ensures that all
components in the realm that implement shutdown (notably targeting
netstack in follow-up change) perform all teardown necessary and ensure
all threads are joined, catching any possible panics in that process.

Note that TestRealm now carries an Arc so we can maintain the
semantics of a TestInterface being sufficient to keep the realm in which
it was created alive.

Original-Bug: 380897722
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1257508
Original-Revision: bceaf39e5702df65579bb8957650359aa3671634
GitOrigin-RevId: 51964ba239398e00135a5c2566b8af47b2b3ac26
Change-Id: Ice3af98a4848f03a49cd2afa0a838de439110c67