[roll] Roll fuchsia [tracing][fidl] Add defer_transfer TraceConfig Option
Currently, in streaming mode, trace manager attempts to write trace data
as it receives it to a socket back to whomever asked it for data. This
is fine when the consumer of the trace in on device, it's just a socket
write. However, in the case of ffx, the other end is on the host and
this triggers a fair amount of overhead sending data over netstack/usb.

This adds a flag which we can propate through trace manager to hold off
on transferring the file (buffer it to disk) until the trace if over.

Original-Bug: 413106095
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260406
Original-Revision: 2e43bd4ccef9cab13dcd2a5dcb3621a9dc9a354f
GitOrigin-RevId: be74b668e7188545ab80019384addfcdd52434aa
Change-Id: If0dde8eca6545234fe9cb43ef47d0862fa2411f0