[roll] Roll fuchsia [magma] Support creating context with priority
Add support for passing a priority to the context.
Creating a context without a priority defaults to medium.

The MSDs are not updated to make use of the priority, they
forward the calls to creating a context without one.

The priority values are based off of VK_EXT_global_priority.

Original-Bug: 402461734
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1243244
Original-Revision: 934fec149968af3b1359ef8e883eed39f7167b01
GitOrigin-RevId: 18a949c78d00e0b8218bdad1d4b8fd63df2c3156
Change-Id: Ifb994852a8b1f6071407979953264c418f2862d6