[roll] Roll fuchsia [netstack3] Fix reassembly of IP fragments
Fix a bug in the reassembly of out-of-order IP fragments. This original
code was incorrectly computing the gaps between fragments, only when 1)
the current fragment fills a gap *before* an existing fragment (i.e. an
out-of-order fragment) and 2) when the current fragment's size is
greater than `FRAGMENT_BLOCK_SIZE`.

The fix itself is trivial; the majority of this diff in this CL is to
update the tests to support testing reassembly on fragments of different
sizes. The original test coverage only exercised fragments with
`FRAGMENT_BLOCK_SIZE`, which kept this bug hidden.

Original-Bug: 413346888
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260604
Original-Revision: 0f7af14bd076d941d0b1479fc61d4b4aab549726
GitOrigin-RevId: ebf7a5b458ca2258cfef59cf2da20d9f8e9d174a
Change-Id: Idd824ba6cfa0929b68325b35f72f8c3f5aeee9ef