[roll] Roll fuchsia [kernel][arm64] Restructure harvest lock dropping
Restructures the harvesting code to only drop the lock if the lock is
actually contended, otherwise reset the entry counter. The top level
function is also rearranged to avoid the use of CallUnlocked, which can
lead to errors with the ConsistencyManager, and avoid using a busy loop
when waiting for mark accessed threads and blocking on the mutex
instead.

Resetting the entry_limit, instead of bouncing out and back into the
harvest loop, reduces the average runtime of harvesting on devices by
up to 40%.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1248544
Original-Revision: 771dd5f2a2d92e7f7f15822988d99e63815148dd
GitOrigin-RevId: e56c30961b7ee71f2c0c2c3dd92dadf9dec26a98
Change-Id: I95cb79427c9ee94c85163d883fabfad7019acb42