[roll] Roll fuchsia [starnix][ext4] Move pager out of starnix kernel
When suspending a Starnix container all the threads in the container are
paused. If the pager threads are suspended before the pager client's the
suspend operation may deadlock. This is because the client threads can
end up blocked in Zircon (on the pager response), which prevents them
from being suspended.

This change moves the pager threads into the starnix_runner component,
which is the same component that is responsible for suspending the
container threads. This guarantees that the pager requests can always be
handled.

Original-Fixed: 410933064

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260704
Original-Revision: 1de8b923df5284ec1007a07862fba86224280cba
GitOrigin-RevId: d35779c922a5d703d32723fc5d4c05390f968478
Change-Id: I57077d6b2f95d376347b1c532789f8a156f21503