[roll] Roll fuchsia [memory] Factor out testing code.
It is not possible to depend on the code gated by `#[cfg(test)]`
from another crate. Having this code in a separate crate would
require to define the trait in one crate, the testing code in
another, and have memory/attribution/processing depend on both.

This solution is simpler, and the struct will be stripped out
by the linker.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1262368
Original-Revision: 241c5d50ec3c8b00374ccb39d9154333409511d1
GitOrigin-RevId: 43e4000a1a674eadce06663f1d12665848b9dd35
Change-Id: I2685d20902e78d874dca2908b736c490fcc3319a