[roll] Roll fuchsia [audio][adr] Handle Composite removals
When a Composite device is removed, a hanging WatchTopology
call completes with an error. This fixes the ADR so that in
that case it does not read into the result (thinking the
call was successful). Also, correctly interpret a subsequent
PEER_CLOSED from signalprocessing as device removal rather
than device error, in the device management state machine.

Adjust signalprocessing logging for clarity; fix clang-tidy
issues in those files while we are here.

Original-Bug: 392662238

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261445
Original-Revision: 01c62c9094c15438651ca5f5aec81471c98f44a6
GitOrigin-RevId: 4fcf93e9131d5d755d3fbfe981c113ae1948c518
Change-Id: I0e55d23d6a77460633bbb8a0f2e8800ea634e04a