[roll] Roll fuchsia [kernel][vm] Fully initialize VMO pages
The PageQueues expect, via ASSERTs, that the initial object and
page_offset are cleared. Not initializing this has happened to work so
far due to these being zeroed at kernel boot, and always reset to zero
being the page is returned.

However, as the vm_page_t::object field is a union, other page states
that do not clear their data before returning the page can result in a
freshly allocated vm_page_t having non-zero object or page_offset
fields which, due to not being cleared, cause the asserts in the
PageQueues to spuriously fail.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1262185
Original-Revision: 2fb9003b747863e5644933a7bf93990198c42634
GitOrigin-RevId: b7ea55318720188ffcd2fb192c450dc0be043f6d
Change-Id: I3ed8ffdbb25e945c6d3b70f4a1b2e45140036dcb