[roll] Roll fuchsia [fxfs] Use Arc::new_cyclic with the pager
PagerPacketReceiver required 2 phase construction because the receiver
needs a Weak of the file and file needs to own the receiver.

Arc::new_cyclic was stabilized a while ago which gives out a Weak to the
file before the file is constructed. This allows the receiver to be
constructed without needing an uninitialized state and to be initialized
later.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261164
Original-Revision: c5dd813e020d9abed074c33039e3484f88e32f76
GitOrigin-RevId: 917d2dbbbca85ed4c18e4481d09ebf7edf0be8b4
Change-Id: Icf61dd5892de547f5fbf04af9ced3ed8754b4ecc