[roll] Roll fuchsia [fidl][ffx] RemoteControl is server=platform
Mark the RemoteControl protocol as server=platform to allow easier
evolution. This allows us to add the host_tool category of FIDL to the
abi checker. The definition of "platform" in the host tool context is
weird and ambiguous but the remote control service is definitely in
it...

Original-Bug: 413661651
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261484
Original-Revision: 5e86b76f129a2d010369928e81d85776560e2e4b
GitOrigin-RevId: 8638fead1614b496f89cd544ebe34cfd69b1e022
Change-Id: I4937ee941dc69aa9d95aae89336f90b89e036acd