[roll] Roll fuchsia [netstack3] Bake ICMP mutability into trait
Now that there are conntrack entries associated with ICMP errors, NAT
will attempt to rewrite the ID of the outer ICMP message if it's an
error and there's a valid inner payload.  Previously, the code assumed
this was impossible, so update_icmp_id will panic.

To resolve the issue, this change adds an IS_REWRITABLE constant to
the IcmpMessage trait that can be checked before the code attempts to
rewrite the ID.

Original-Bug: 328057704
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1262884
Original-Revision: 00edafa4e0bbf0c1d50278e59271e3575e5aea3f
GitOrigin-RevId: f7a4e2a50c8100c2e292f7bbacddd412691aca54
Change-Id: Ibebc415dc98093df4963d51b640565821cad4feb