[roll] Roll fuchsia [fidl][fuchsia.starnix.binder] Add ProcessAccessor::WriteBytes
This allows us to write the memory of the server process without
creating a VMO.

To create a VMO takes roughly 1KiB of memory, and writing to it takes at
minimum an additional 4KiB of memory. The time it takes to create and
copy data into the VMO before passing it across the channel is also
higher than simply writing bytes into the message.

To short-circuit all of this for small writes, we are adding WriteBytes.

Original-Bug: 279766373
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1259684
Original-Revision: ff2a28496e73da066f3e375ee40180a3f142dca4
GitOrigin-RevId: befc6a462d1aef1cadf1292146f869338a1d5381
Change-Id: I401d9bd73b5905a2a4feafb8b1a6e78a6ccf5dbc