[roll] Roll fuchsia [ffx] Use scope ID strings for ssh.
Scope ID here refers to a link-local IPv6 address scope ID.

Instead of running ssh commands with scope ID's as integers, use
verified scope ID strings.

This adds a `ScopedSocketAddr` struct to be used in ssh and related
commands. This structure is intended to format a stable scope ID in the
form of a string. It also, when constructed, will verify the
scope ID of the address, returning an error if it cannot be found.

The approach taken here is to prevent having to rewrite the majority
of ffx daemon related code. It is not necessarily the most complete.
Inside the daemon is several layers (in several directions) of
address structs: TargetAddrInfo, TargetIp, TargetIpPort, TargetInfo,
TargetAddr, etc. Some of these are directly useful as they abstract
things like vsock connections, but most have dubious usefulness,
especially with the advent of moving more functionality away from the
daemon. Ideally this `ScopedSocketAddr` would be pushed closer and
closer to the first invocation of the user, therefore preventing invalid
scope ID's from making it to the `ssh` command invocation.

Original-Bug: 411230336

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260729
Original-Revision: d21abea81e7eb12008b45e4a29291ca2a2d6bb02
GitOrigin-RevId: 65ed52ed1621480240681b8b06f2b946357f9c3e
Change-Id: Iff9c7bc00118a5f8a4f95d868cad5044dc3ead99