[roll] Roll fuchsia [tracing] Rename instant_flow_* to instaflow_*
Matches the newly-added C++ macro names.  Since the C++ names are part
of the public SDK, we decided that it was better to pick the name that
we preferred (since it would be difficult to change later), and make the
Rust macros match.

Original-Fixed: 413095076
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260824
Original-Revision: c88d744315f88f562a975d993643e7b76a2758ba
GitOrigin-RevId: 4fef2d8b5958190091ee4891c748cc80b4103377
Change-Id: I2c3c8db5d239668405cdcf06c986556aab500eaa