[roll] Roll fuchsia [fxfs] Fix profile tests. Shouldn't be marking as purged.
These were being marked as purged to receive an open count and verify
that it was zero. Not only was this not a sure way to check, but marking
an object to be purged without adding it to the graveyard isn't safe if
the purge actually happens, it can result in a partially deleted object.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1263964
Original-Revision: 26bd04eecf12de5ca248e4e4a43b7d1aa3cdb87b
GitOrigin-RevId: 56fc39bc66390fa793ae523bfec7fa7dd762d5a8
Change-Id: Id677d7127e1c3e5d3fcff6a62b5d01c05f63a63b