[roll] Roll fuchsia [io] Workaround for remove-and-rename.
fuchsia.io.* related Deprecated methods aren't quite correct.

For example this:

    @available(added=25, removed=26)
    @selector("Open")
    strict DeprecatedOpen(resource struct {

This will keep DeprecatedOpen around if and only if API level 25 is
supported, but that's not what we want. We want it to be supported (for
compatibility purposes) as long as any API level that supported Open is
supported (that is, any between 11 and 25).

We don't have a nice way of doing that right now. This CL changes them
to work in a variety of _different_ bad ways, but ways that allow us to
retire unused API levels without things breaking.

Original-Bug: 324080864
Original-Bug: 384994764

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1255651
Original-Revision: 6ae79525dd87ecdf5bc86d94366cd2821c6e1dda
GitOrigin-RevId: 5a44ba6a403a3c4ca05cc3f06e06cdd649a4d348
Change-Id: Ic00375d221e57a1d467e056f183133359cefac46