[roll] Roll fuchsia [netstack3/filter] Add ICMP payload modification
This is the bulk of the mechanism required to support NAT for ICMP
errors, where we want to be able to rewrite the inner IP and transport
payloads in addition to the outer IP header.

The basic idea of the implementation is that we return something that
implements IpPacket on the inner payload to allow modification just like
normal IP and transport headers.

Also catch a few more places where FilterIpExt needed to be plumbed.

Original-Bug: 328057704
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1255148
Original-Revision: 33f3ea7e4c822f9d897698102373b87be4b439a9
GitOrigin-RevId: 261cf6641f8f00c11ef292302f062386c642a90e
Change-Id: I4babb3133c4b81bdec53b63685284a7062087012