[roll] Roll fuchsia [block_server] Split large requests
Split large requests in the "core" block server before they make it to
the concrete block server implementation.

Large requests are placed into an ad-hoc "one-shot" group, to ensure
that only one reply goes out on the FIFO.  (For requests that are
already in a group, we simply accumulate the additional requests into
the existing group).

I opted to split requests at the same time that we check the OffsetMap
for remapping requests, because we will also need to split requests for
discontiguous requests.

Also adds the ability to set maximum_transfer_size to the ramdisk, and
fixes a bug (TRIM sends two responses).  This is used for a C++
integration test.

Original-Fixed: 372555590
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261608
Original-Revision: f3b66b42762a99dddaa476ef438567e015296698
GitOrigin-RevId: 53ac91960432dfc12232f26b03052d59a4626805
Change-Id: I086b117a3cef1ed0f9fd21c045d4deee27240b57