[roll] Roll fuchsia [logs] Remove explicit shard reference in RealmBuilder shard
The RealmBuilder shard is shipped in the SDK.
SDK consumers may want custom routing for LogSink
and InspectSink. Forcing them to include the shard
removes a lot of flexibility, and makes migrations
of LogSink routing more difficult, not simpler.
All out-of-tree users have migrated to explicit
routes or explicit dependencies on the LogSink shard,
so it's safe to update this.

Original-Fixed: 404543928
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1258026
Original-Revision: 808411b3edeedbae97ca763a1f11d158356d07ce
GitOrigin-RevId: 2b4a7d240641ee496ff2d9f7c5cb5db298266b8b
Change-Id: I6a01fd29ce9a63e51e4c66c36e7c80bf200986d1