[roll] Roll infra/recipes [milestone_roller] Upload CLs even if config generation fails
Config generation is currently failing on this builder due to a race
condition where a CQ-blocking builder was deleted since the new branch's
base commit was created:
https://screenshot.googleplex.com/62p8yUM6CZPUvsF

Currently the only ways to fix this situation are to manually run the
steps that the recipe runs and fix any issues, or to iteratively fix
issues raised by the roller until it's happy.

If we have the roller upload a CL even if it's going to fail CQ, a human
can go in and see why it's failing and amend the CL to make it
submittable, which is a much easier process.

Original-Bug: b/414618041
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1264764
Original-Revision: f010c755cb02c7eab87a215b9393d1057858577c
GitOrigin-RevId: f741d0dee29e1b8b31e3299cf13a43cfb288bebf
Change-Id: I95987e14ea1adca0d163e0deef665d16a35814f3