[roll] Roll fuchsia [time][tests] Convert some tests to run in fake time
Quite a few of timekeeper tests have been written to run in real time.
Unfortunately, this means that the tests are sensitive to adding more
coroutine work, or changing the task scheduling. Both lead to flakes.

I noticed a few tests that were failing this way locally when I fixed
the way initial sampling delay is treated, and I converted those tests
to run in fake time.  The main difference is that we no longer run until
the event loop is stalled, which is a fickle criterion. Instead, we run
the event loop for a predefined length of (fake) time.

This fix is not directly related to b/404936065, but should make the
upcoming changes test more predictably.

Original-Bug: 404936065
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1263826
Original-Revision: 2a4e51e201ec344d4955e8dd73d953da6d290e38
GitOrigin-RevId: 5f2e8d96d490fd3c254ff5ad70df7b4dee5cb27c
Change-Id: I20759a3916cb94e58ffc85afda4503e0eaa79f65