[roll] Roll fuchsia [audio][test] Make VirtualAudioDriver compliant
VirtualAudioComposite instances are now detectable by the
AudioDeviceRegistry service, but ADR reveals that the
default configuration for Composite does not comply with
requirements for the `signalprocessing` protocol.

By default VirtualAudioComposite exposes a RingBuffer routed
to a DAI. The DAI does not set `plug_detect_capabilities` as
required, and in response to `WatchElementState` neither of
the two elements specify the required field `started`.

This CL fixes those omissions, making the default config
compliant. This is a prerequisite for VAD to be used as a
test fixture, to hermetically test ADR and its clients.

We also update virtual_audio_util so that 'Composite' is the
default assumed driver type. While we are here, clean up any
clang-tidy issues in this source file.

Original-Bug: 392662238

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261446
Original-Revision: 55a733e0d5a384f603ed4c991460e7e84eedf26a
GitOrigin-RevId: 22708c9f7c41b4cba209a0ffe27a9a06c4e72bc7
Change-Id: I1959217f7c9881d643d0f4f1a35752a8a6db8b94