[roll] Roll fuchsia [botanist] Do not powercyle NUCs in fastboot during retry.
Powercycling NUCs in fastboot will cause the iPXE chain loader to run
and take the devices out of fastboot momentarily. Also, the check for
presence of `USB RESET` in serial logs after powercylcing will fail
because this is not part of the logs when iPXE is booting the NUCs into
fastboot.

We can rely on DMS to get the NUCs into fastboot. Botanist should simply
retry to flash or boot the NUC.

Original-Bug: 346868967
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1265284
Original-Revision: e661d38c3ef18ebd6ddb805017aecf7f7062bb7d
GitOrigin-RevId: 36f97d419c0869bcda85ef2e391018f6c2d6ade5
Change-Id: I953cd6cced4176a53985cc6cdb178dcc69059d52