[roll] Roll fuchsia [component_manager] use the route target when logging routing errors
We were using the component at which a route failed to route errors
about the logging, whereas we want to use the component that started the
routing operation.

This change accomplishes that by relying on the target in the route's
metadata when an error occurs.

Original-Bug: 411475421
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1262406
Original-Revision: ac22e9062beea51045093d2e841f4b3097eb9510
GitOrigin-RevId: 6ab19abed0eb5347c38e4753ba47bd487f146083
Change-Id: I680dc4aef0d5ac23f8ea1e268b71253758169ccc