[roll] Roll fuchsia [time] Fix user time adjust test
I realized that this test is sensitive to the discrepancy between
monotonic timeline and boot timeline. This leads to different test
results on devices that spent different amount of time suspended.

This change removes the issue for this particular test. But there
are other tests with the same problem, which also need fixing. I
am leaving them for a separate change.

Original-Bug: 404936065
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1265231
Original-Revision: d1569cf93034038b3950ba58920b65cc28faba56
GitOrigin-RevId: 6e6f3b5143cfe17a2f2f36caf0624586ecf84e1c
Change-Id: I9322a283fa64ce3e146f373d8d412dbfd55fe791