[roll] Roll fuchsia [fxfs] Support allow_existing in BlobCreator.
The allow_existing flag will not fail the request if the object already
exists. This also changes edge case behaviour where two writers that
start writing together will both see success when they finish rather
than the second receiving a failure on the final write.

Swaps out the backing for the serving vmo when a blob is replaced while
open. This is done by using an Arc on the registration so that it can be
attached to both the old and the new blob before moving the ownership to
the new blob by pointing the registration at it and adjusting the open
counts to reflect that it has taken ownership of the vmo.

Original-Bug: 409839447

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1263965
Original-Revision: 03f329348a91591715479090d00c7a38f23dee98
GitOrigin-RevId: 7948cc03737eb7595d4511c34123b33741a368ff
Change-Id: I6fc8c104308c22156bc72e05978d63eb9b54ff39