[roll] Roll fuchsia [ffx] Don't break namespace querying for FDomain over daemon
Calling the remote factory is supposed to do a bunch of retrying and
re-establishing for the daemon. In preserving this behavior for the
FDomain version of the remote factory we broke calling .domain() on the
proxies we got from the factory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1251023
Original-Revision: 6c6d6c6b3ba0e77c80acb4760ca4353fc35c3535
GitOrigin-RevId: 9cb2fccdb8b7f593eaf0e0a7e8b698b98b572dfe
Change-Id: I1bccea72fc6233ce8fe71df71e721aafd9f3824e