[roll] Roll fuchsia [network/testing] update expectations to flaky
The network conformance case 7.4 in the ip/ip_router suites verifies
that the fragmentation cache sets timeouts based on the TTL of the
fragment. This is currently not supported in NS3.

This ultimately leads to a race between the test sending the final
fragment, and the netstack timing out the existing fragments in the
cache. Once we implement this feature the race should be gone.

Original-Bug: 414413500
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1266664
Original-Revision: fcbdbd16f25dfe00342f2aa05a1d7b056ddf845d
GitOrigin-RevId: 450917921fc394d8bcc8c8d59cb013ddf751742e
Change-Id: I4a473e1a61b3ce24946fdb116062795361a36324