[roll] Roll fuchsia [fuchsia-component] Add helpers for custom service name
This change introduces helpers for components that connect to services
that have been given a custom name, and so cannot use the service marker
to auto-fill the desired capability name.

These helpers are to be used, for example, with the proprietary TEE
service implementation in which a service with a proprietary name is
routed through the component topology under a platform-defined name.

Original-Bug: b/410521972
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1266744
Original-Revision: 29c811164599947170bcb540162b1132693c516b
GitOrigin-RevId: b6ed287f0671b8b82ee767461ad47e0e7c0c1321
Change-Id: Id13c53895913746f318b576c84556bc23da01fe4