[roll] Roll fuchsia [netstack3] Merge IPv4/IPv6 Address state
In preparation for supporting DAD on IPv4, use the same types to
represent the state of an address on a device. `Ipv4AddressEntry` and
`Ipv6AddressEntry` are now more alike then they are different.

In doing so, I can remove a handful of traits that were used to
trampoline types around.

Original-Bug: 42077260

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1266824
Original-Revision: 152688a9534c923869763cfc171b3f28923e3a27
GitOrigin-RevId: 43617117bbb355236bd577b4c424e65ffb5f5c5e
Change-Id: Ida3634331ed79d3cc3ec31585d3d07d550f09b3d