[roll] Roll fuchsia [netstack3] Limit default route interface watching to main table
Only update has_default_route in the interfaces watcher if we are adding
or removing a route to the main table.

This avoids a bug where we indicate that there is no longer a default
route at all when in fact we've merely removed the last default route in
the table (while other default routes are installed in other tables).

Original-Bug: 412695202
Multiply: interfaces-fidl
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1266945
Original-Revision: 4d6bf34d50ed084a52e5dc638e28129165cb1cf6
GitOrigin-RevId: 0423e792b6e00f07e85e1ea94961408e4ba7fc05
Change-Id: I92440983e53c84b2e0d871361fbd9624462c5e4e