[roll] Roll fuchsia [fidl][rust][fdomain] Build rule cleanup
This is a clean-up of how we build the flavours of Rust FIDL
bindings in our glorious FDomain world. It moves all of the logic and
policy (other than a little backwards compat stuff) from
//build/fidl/fidl.gni to //build/rust/fidl_rust.gni where it belongs.

As a side effect this makes the "common" create for FIDL libraries
depend on the "common" crate for their dependencies rather than on the
main end-user FIDL bindings crate. I noticed that these were showing up
in build trace critical paths. In a local test it's cutting a couple of
seconds off the hot path for a clean ffx build.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260905
Original-Revision: 38f683d82c98fdffe38a70933c76d4ad67331cca
GitOrigin-RevId: 6417f54c2b605e71701f118bc1512e47604b93ca
Change-Id: I217873973342fcdcf7e149147c4f9ab0784aeb95