[roll] Roll fuchsia [netstack3] Implement ICMP error NAT
This relies on the ICMP error modification code that was previously
added to the packets module. When we see a valid inner payload, we
perform inverse NAT on it before rewriting the outer IP header.

For a nitty-gritty justification of the implementation, see the comment
in the rewrite_icmp_error_payload function.

Original-Fixed: 328057704
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1263924
Original-Revision: a33f80846ed0ca0be8e051e7c68ece476898b090
GitOrigin-RevId: 598132496aa7751ebc292d41b8ff59aa890f2a93
Change-Id: I9fd993e3e079e3b1272a0ff91985fcbee3453d40